requirements.txt
openai==1.42.0
python-dotenv==1.0.1
PyPDF==4.3.0
langchain==0.2.14
langchain-community == 0.2.12
sentence-transformers==3.0.1
chromadb==0.5.3
langchain_chroma==0.1.2
langchain_openai==0.1.16
beautifulsoup4==4.12.3
gradio==4.41.0
langchain-huggingface==0.0.3
